requirements.txt
addict==2.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.2.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.3
cycler==0.12.1
dash==4.0.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
easydict==1.13
embreex==2.17.7.post7
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
Flask==3.1.3
fonttools==4.61.1
fqdn==1.5.1
freetype-py==2.5.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
human_body_prior @ git+https://github.com/nghorbani/human_body_prior@78c86eae5ed518ae22bf197fd74211bbfa45551a
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
loguru==0.7.3
lxml==6.0.2
manifold3d==3.4.0
mapbox_earcut==2.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mistune==3.2.0
mpmath==1.3.0
narwhals==2.17.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook_shim==0.2.4
numpy==2.4.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open3d==0.19.0
overrides==7.7.0
packaging @ file:///home/task_176104885106445/conda-bld/packaging_1761049078006/work
pandas==3.0.1
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.9.2
plotly==6.5.2
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycollada==0.9.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyglet==1.5.31
Pygments==2.19.2
PyOpenGL==3.1.0
pyparsing==3.3.2
pyquaternion==0.9.9
pyrender==0.1.45
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
PyVirtualDisplay==3.0
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
rtree==1.4.1
scikit-learn==1.8.0
scipy==1.17.1
Send2Trash==2.1.0
sentry-sdk==2.54.0
shapely==2.1.2
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
stack-data==0.6.3
svg.path==7.0
sympy==1.13.1
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transforms3d==0.4.2
trimesh==4.11.2
triton==3.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
vhacdx==0.0.10
wandb==0.25.1
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.6
widgetsnbextension==4.0.15
xxhash==3.6.0
yacs==0.1.8
zipp==3.23.0
